% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{checkIterArgs}
\alias{checkIterArgs}
\title{Used to inherit roxygen docs}
\usage{
checkIterArgs(gradtol, steptol, iterlim)
}
\arguments{
\item{gradtol}{Parameter for the nlm optimisation - a positive scalar giving the tolerance at which the scaled gradient is considered close enough to zero to terminate the algorithm (see [\code{nlm} doccumentation](https://stat.ethz.ch/R-manual/R-patched/library/stats/html/nlm.html)).}

\item{steptol}{Parameter for the nlm optimisation - a positive scalar providing the minimum allowable relative step length (see [\code{nlm} doccumentation](https://stat.ethz.ch/R-manual/R-patched/library/stats/html/nlm.html)).}

\item{iterlim}{Parameter for the nlm optimisation - a positive integer specifying the maximum number of iterations to be performed before the program is terminated (see [\code{nlm} doccumentation](https://stat.ethz.ch/R-manual/R-patched/library/stats/html/nlm.html)).}
}
\description{
Used to inherit roxygen docs
}
\keyword{internal}
