% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmatch.R
\name{dmatching}
\alias{dmatching}
\title{Mass function of the generalised matching distribution}
\usage{
dmatching(x, size, prob = 0, log = FALSE)
}
\arguments{
\item{x}{A vector of numeric values to be used as arguments for the mass function}

\item{size}{The size parameter for the generalised matching distribution (number of objects to match)}

\item{prob}{The probability parameter for the generalised matching distribution (probability of known match)}

\item{log}{A logical value specifying whether results should be returned as log-probabilities}
}
\value{
If all inputs are correctly specified (i.e., parameters are in allowable range) then the output will
be a vector of probabilities/log-probabilities corresponding to the vector argument x
}
\description{
\code{dmatch} returns the probability or log-probability values for the arguments.
}
\details{
This function computes probabilities or log-probabilities from the mass function of the generalised matching
distribution.  Further details on the distribution can be found in the following paper:

O'Neill, B. (2021) A generalised matching distribution for the problem of coincidences.
}
