% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmatch.R
\name{pmatch}
\alias{pmatch}
\title{Cumulative distribution function of the generalised matching distribution}
\usage{
pmatch(q, size, prob = 0, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{A vector of numeric values to be used as arguments for the cumulative distribution function}

\item{size}{The size parameter for the generalised matching distribution (number of objects to match)}

\item{prob}{The probability parameter for the generalised matching distribution (probability of known match)}

\item{lower.tail}{A logical value specifying whether results should be returned for the lower or upper tail}

\item{log.p}{A logical value specifying whether results should be returned as log-probabilities}
}
\value{
If all inputs are correctly specified (i.e., parameters are in allowable range) then the output will
be a vector of probabilities/log-probabilities corresponding to the vector argument x
}
\description{
\code{pmatch} returns the probability or log-probability values for the arguments.
}
\details{
This function computes probabilities or log-probabilities from the cumulative distribution function of the 
generalised matching distribution.  Further details on the distribution can be found in the following paper:

O'Neill, B. (2021) A generalised matching distribution for the problem of coincidences.
}
