% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatch.R
\name{qmatch}
\alias{qmatch}
\title{Quantile function of the generalised matching distribution}
\usage{
qmatch(p, size, prob = 0, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{A vector of numeric probability/log-probability values to be used as arguments for the quantile function}

\item{size}{The size parameter for the generalised matching distribution (number of objects to match)}

\item{prob}{The probability parameter for the generalised matching distribution (probability of known match)}

\item{lower.tail}{A logical value specifying whether the input represents lower or upper tail probabilities}

\item{log.p}{A logical value specifying whether the input values in p are log-probabilities}
}
\value{
If all inputs are correctly specified (i.e., parameters are in allowable range) then the output will
be a vector of quantiles corresponding to the vector argument p
}
\description{
\code{qmatch} returns the quantiles for the arguments.
}
\details{
This function computes quantiles from the quantile function of the generalised matching distribution.  Further 
details on the distribution can be found in the following paper:

O'Neill, B. (2021) A generalised matching distribution for the problem of coincidences.
}
