% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmatch.R
\name{rmatch}
\alias{rmatch}
\title{Random generation from the generalised matching distribution}
\usage{
rmatch(n, size, prob = 0)
}
\arguments{
\item{n}{A non-negative integer specifying the number of rdandom values to produce}

\item{size}{The size parameter for the generalised matching distribution (number of objects to match)}

\item{prob}{The probability parameter for the generalised matching distribution (probability of known match)}
}
\value{
If all inputs are correctly specified (i.e., parameters are in allowable range) then the output will
be a vector of pseudo-random numbers from the distribution
}
\description{
\code{rmatch} returns the a vector of pseudo-random values from the distribution.
}
\details{
This function generates pseudo-random values from the generalised matching distribution.  Further details on 
the distribution can be found in the following paper:

O'Neill, B. (2021) A generalised matching distribution for the problem of coincidences.
}
